+++ src/main/java/com/project/chamjimayo/domain/entity/Comment.java
@@ -23,7 +23,7 @@ public class Comment extends BaseEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "comment_id")
-	private Integer commentId;
+	private Long commentId;
 
 	// 게시글 아이디 (댓글이 달린 게시글)
 	@ManyToOne(fetch = FetchType.LAZY)
@@ -42,6 +42,5 @@ public class Comment extends BaseEntity {
 	// 추천 수
 	@Column(name = "like_count")
 	private Integer likeCount;
-
 }
